refactor(regex): split header properties and tests
(cherry picked from commit c217a231c0e0133885ce105aa66de313f6c70be6)
